Add files via upload
Implementation details
Sensor Inputs
Sensor inputs are simulated using three switches. I have used SPST switches (with
latched action)
When the switch is open, the output voltage is 5V (high).
Since an output voltage of 5V is needed (high level logic) when the switch is closed,
hence I’ve used an inverter at the switch output.
Priority resolving circuit
When Switch 1 is closed, the priority resolving circuit output has to be 1, regardless
of other switches. Hence Switch 1 output itself can be used.
When Switch 2 is closed, the priority resolving circuit output for Switch 2 has to be 1,
subject to Switch 1 being open. Hence, a 2-input AND gate is used to implement
this.
When Switch 3 is closed, the priority resolving circuit output for switch 3 has to be 1,
subject to both Switch 1 and Switch 2 being open. This is implemented using a 3-
input AND gate
DEMUX Circuit
There are three outputs from the Priority resolving circuit but six outputs are needed
from the DEMUX.
These six outputs are PRESET_1, CLEAR_1, PRESET_2, CLEAR_2, PRESET_3
and CLEAR_3.
Whenever PRESET_1 is 1, CLEAR_2 and CLEAR_3 should also be 1.
Similarly, when PRESET_2 is 1, CLEAR_1 and CLEAR_3 should be 1 and
When PRESET_3 is 1, CLEAR_1 and CLEAR_2 should be 1.
This is to make sure that only one of ‘1’, ‘2’ or ‘3’ is 1, not more.
Since all these outputs are active low, there are inverters in the path of all of them.
1 hot counter logic
3 D flip-flops are used that have PRESET and CLEAR (active low) inputs.
Clock is a 5 sec period clock from clock generator.
All flip-flops start at 0 output at the start of the simulation.
So, there is a logic at the D input of ‘1’ flip-flop which checks for this condition to
make it to 1 at the start of the next clock.
‘2’ is also PRESET to 1 when ‘3’ flip-flop is 1.
‘2’ and ‘3’ flip-flops are just shifted versions of ‘1’ and hence they are just serially
connected.
Green LEDs are connected to Q outputs and Red LEDs are connected to Qbar
outputs of the D flip flops. (This also ensures that Red and LED lights of the same
road are never ON together)
Simulation
Tested with all switches open and then with one or more switches closed in random
fashion and checked the output.
The LED lights are as per expectation.
At no point more than one Green LED was seen to be active.